Apply a function over spatially chunked data

Split a function operation into multiple spatial chunks. Results are appended back into the database as a new table. This is not parallelized as some databases do not work with parallel writes, but are more performant with large single chunks of data. The functions that are provided, however, can be parallelized in their processing after the chunk has been pulled into memory and only needs to be combined into one before being written.

chunkSpatApplyPoly(
  x = NULL,
  y = NULL,
  chunk_y = TRUE,
  fun,
  extent = NULL,
  n_per_chunk = getOption("gdb.nperchunk", 1e+05),
  remote_name = NULL,
  progress = TRUE,
  ...
)

chunkSpatApplyPoints(
  x = NULL,
  y = NULL,
  chunk_y = TRUE,
  fun,
  extent = NULL,
  n_per_chunk = getOption("gdb.nperchunk", 1e+05),
  remote_name = NULL,
  progress = TRUE
)

Arguments

x

dbPolygonProxy or dbPointsProxy

y

missing/null if not needed. Otherwise accepts a dbPolygonProxy or dbPointsProxy object

chunk_y

(default = TRUE) whether y also needs to be spatially chunked if it is provided.

fun

function to apply

extent

spatial extent of data to apply across. Defaults to the extent of x if not given

n_per_chunk

(default is 1e5) number of records to try to process per chunk. This value can be set globally using options(gdb.nperchunk = ?)

remote_name

name to assign the result in the database. Defaults to a generic incrementing 'gdb_nnn' if not given

progress

whether to plot the progress

...

additional params to pass to Duckling object creation

Value

dbPolygonProxy or dbPointsProxy