Create a dense matrix with database backend

Create an S4 dbMatrix object that has a dense ijx triplet vector format (ijx). The data for the matrix is either written to a specified database file or could also be read in from files on disk in a chunked manner.

createDBDenseMatrix(
  matrix,
  remote_name = "mat_test",
  db_path = ":temp:",
  overwrite = FALSE,
  cores = 1L,
  nlines = 10000L,
  callback = callback_formatIJX(),
  custom_table_fields = fields_preset$dbDenseMatrix,
  dims,
  dim_names,
  ...
)

Arguments

matrix

object coercible to matrix or filepath to matrix data accessible by one of the read functions. Can also be a pre-prepared tbl_sql to compatible database table

remote_name

name to assign within database

db_path

path to database on disk

overwrite

whether to overwrite if table already exists in database

cores

number of cores to use if reading into database

nlines

number of lines to read per chunk if reading into database

callback

callback functions to apply to each data chunk before it is sent to the database backend

dims

dimensions of the matrix (optional)

dim_names

list of rownames and colnames of the matrix (optional)

...

additional params to pass

Details

Information is only read into the database during this process. Based on the remote_name and db_path a lazy connection is then made downstream during dbData initialization and appended to the object. If a dplyr tbl is provided as pre-made input then it is evaluated for whether it is a tbl_Pool and whether the table exists within the specified backend then directly passed downstream.