Create a framework for chunked processing with terra polygon SpatVectors

Create an S4 dbPolygonProxy object that is composed of two database tables. One table hold geometry information while the other holds attribute information.

createDBPolygonProxy(
  SpatVector,
  remote_name = "poly_test",
  db_path = ":temp:",
  id_col = "poly_ID",
  xy_col = c("x", "y"),
  extent = NULL,
  overwrite = FALSE,
  chunk_size = 10000L,
  callback = NULL,
  custom_table_fields = fields_preset$dbPoly_geom,
  custom_table_fields_attr = NULL,
  attributes = NULL,
  ...
)

Arguments

SpatVector

object coercible to SpatVector or filepath to spatial data readable by vect

remote_name

name of remote table on database backend

db_path

filepath to the database backend

id_col

column in data to read in that contains the polygon id information

xy_col

columns in data to read in that contain the x and y vertex info

extent

terra SpatExtent (optional) that can be used to subset the data to read in before it is saved to database

overwrite

whether to overwrite if remote_name already exists on the database

chunk_size

the number of polygons to read in per chunk read

callback

data formatting and manipulations to perform chunkwise before the data is saved to database. Instructions should be provided as a function that takes an input of a data.table and returns a data.table

custom_table_fields

(optional) custom table field SQL settings to use during table creation for the spatial geometry table

custom_table_fields_attr

(optional) custom table field SQL settings to use during table creation for the attributes table

attributes

(optional) a tbl_sql connected to the database backend that contains the associated attributes table. Only used if a pre-made tbl_sql is also provided to SpatVector param

Details

Information is only read into the database during this process. Based on the remote_name and db_path a lazy connection is then made downstream during dbData intialization and appended to the object. If the data already exists within the database backend then it is entirely permissible to omit the SpatVector param.