Represent polygon as single xy coordinate pair stand-ins

Internal. Calculate simple representative spatial locations for polygons. Depending on method param, finds the polygon vertex that is the mean or at the lower left, lower right, upper left, or upper right. All methods are roughly equivalent in speed. Note that this is NOT the polygon centroid.

simplify_poly_as_point(
  x,
  method = c("mean", "lowL", "lowR", "upL", "upR"),
  ...
)

Arguments

x

tbl_Pool (usually from a dbPolygonProxy object)

method

one of 'mean', 'lowL', 'lowR', 'upL', 'upR'

...

additional params to pass

Value

tbl_Pool of xy coordinates

Details

Generate a representative point for all the polygons in a dbPolygonProxy. This facilitates spatial selection of the polygons as there are fewer points to keep track of, making it less computationally intensive and less ambiguous whether a polygon should be selected by a specific selection if parts of it fall in multiple selections.