Stream large flat files to database backend using fread

Files are read in chunks of lines via fread and then converted to the required formatting with plugin functions provided through the callback param before being written/appended to the database table. This is slower than directly writing the information in, but is a scalable approach as it never requires the full dataset to be in memory.
If more than one or a custom callback is needed for the formatting then a combined or new function can be defined on the spot as long as it accepts data.table input and returns a data.table.

streamToDB_fread(
  path,
  backend_ID,
  remote_name = "test",
  idx_col = NULL,
  pk = NULL,
  nlines = 10000L,
  cores = 1L,
  callback = NULL,
  overwrite = FALSE,
  custom_table_fields = NULL,
  ...
)

Arguments

path

path to the matrix file

backend_ID

ID of the backend to use

remote_name

name to assign table of read in values in database

idx_col

character. If not NULL, the specified column will be generated as unique ascending integers

pk

character. Which columns to use as primary key

nlines

integer. Number of lines to read per chunk

cores

fread cores to use

callback

callback functions to apply to each data chunk before it is sent to the database backend

overwrite

whether to overwrite if table already exists (default = FALSE)

custom_table_fields

default = NULL. If desired, custom setup the fields of the table. See dbCreateTable

...

additional params to pass to fread